+++ src/index.ts
@@ -1,5 +1,3 @@
-import 'react-native-get-random-values';
-
 export * from './CoinflowWithdraw';
 export * from './CoinflowPurchase';
 export * from './CoinflowPurchaseHistory';